requirements.txt
# File Content
boto3>=1.18
pytest>=6.2
lorem
sphinx
pydantic
sphinx_rtd_theme
sphinx_autodoc_typehints
nbsphinx
ipython
datetime
uuid
python-magic
chardet
mkdocs
mkdocs-gen-files
mkdocs-material
mkdocs-gen-files 
mkdocs-literate-nav 
mkdocs-section-index
mkdocstrings-crystal